split script.dart into several files for maintainability
